AdminCommand/ConsoleCommand ValidateAccess: (sbyte) cast GetLevel() arg to player.HasAccess(sbyte) (faithful Java byte=sbyte accessLevel); clears 2 CS1503 (901->899)

+++ dotnetConversion/src/Aion.GameServer/Utils/ChatHandlers/AdminCommand.cs
@@ -25,7 +25,7 @@ public abstract class AdminCommand : ChatCommand
 
     public override bool ValidateAccess(Player player)
     {
-        bool hasAccess = player.HasAccess(GetLevel()) || CommandsAccessService.HasAccess(player.GetObjectId(), GetAlias());
+        bool hasAccess = player.HasAccess((sbyte)GetLevel()) || CommandsAccessService.HasAccess(player.GetObjectId(), GetAlias());
         if (!hasAccess && player.IsStaff())
             SendInfo(player, "<You need access level " + GetLevel() + " or higher to use " + GetAliasWithPrefix() + ">");
         return hasAccess;
